moved tests folder to kubernetes dir
