--- cloud/kubernetes/helm/tests/go.mod
@@ -1,9 +1,0 @@
-module github.com/TIBCOSoftware/be-tools/cloud/kubernetes/helm/tests
-
-go 1.14
-
-require (
-	github.com/gruntwork-io/terratest v0.25.1
-	github.com/stretchr/testify v1.4.0
-	k8s.io/api v0.17.0
-)
