--- cloud/kubernetes/helm/tests/template/cluster_cache_store_test.go
@@ -1,108 +1,0 @@
-package template
-
-import (
-	"testing"
-
-	"github.com/TIBCOSoftware/be-tools/cloud/kubernetes/helm/tests/common"
-	"github.com/gruntwork-io/terratest/modules/helm"
-)
-
-func TestAS2CacheRDBMSMysqlStore(t *testing.T) {
-	helmChartPath := "../../"
-	options := &helm.Options{
-		SetValues: common.AS2CacheRDBMSStoreValues(),
-	}
-
-	appAndJmxServices(t, options, helmChartPath)
-
-	// inference agent test
-	inferenceOutput := helm.RenderTemplate(t, options, helmChartPath, "beinferenceagent", []string{"templates/beinferenceagent.yaml"})
-	common.InferenceAS2MysqlTest(inferenceOutput, t)
-
-	// cache agent test
-	cacheAppOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheagent", []string{"templates/becacheagent.yaml"})
-	common.CacheAS2MysqlTest(cacheAppOutput, t)
-
-	// be cache service test
-	beCacheServiceOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheservice", []string{"templates/becache-service.yaml"})
-	common.AS2CacheServiceTest(beCacheServiceOutput, t)
-
-	// configmap test
-	configOutPut := helm.RenderTemplate(t, options, helmChartPath, "configmap", []string{"templates/configmap.yaml"})
-	common.ConfigMapMysqlTest(configOutPut, t)
-}
-
-func TestFTLCacheStoreAS4(t *testing.T) {
-	helmChartPath := "../../"
-	options := &helm.Options{
-		SetValues: common.FTLCacheAS4StoreValues(),
-	}
-
-	appAndJmxServices(t, options, helmChartPath)
-
-	// inference agent test
-	inferenceOutput := helm.RenderTemplate(t, options, helmChartPath, "beinferenceagent", []string{"templates/beinferenceagent.yaml"})
-	common.InferenceFTLAS4Test(inferenceOutput, t)
-
-	// cache agent test
-	cacheAppOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheagent", []string{"templates/becacheagent.yaml"})
-	common.CacheFTLAS4Test(cacheAppOutput, t)
-
-	// be cache service test
-	beCacheServiceOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheservice", []string{"templates/becache-service.yaml"})
-	common.IgniteCacheServiceTest(beCacheServiceOutput, t)
-
-	// configmap test
-	configOutPut := helm.RenderTemplate(t, options, helmChartPath, "configmap", []string{"templates/configmap.yaml"})
-	common.ConfigMapAS4Test(configOutPut, t)
-}
-
-func TestFTLCacheStoreCassandra(t *testing.T) {
-	helmChartPath := "../../"
-	options := &helm.Options{
-		SetValues: common.FTLCacheCassandraStoreValues(),
-	}
-
-	appAndJmxServices(t, options, helmChartPath)
-
-	// inference agent test
-	inferenceOutput := helm.RenderTemplate(t, options, helmChartPath, "beinferenceagent", []string{"templates/beinferenceagent.yaml"})
-	common.InferenceFTLCassTest(inferenceOutput, t)
-
-	// cache agent test
-	cacheAppOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheagent", []string{"templates/becacheagent.yaml"})
-	common.CacheFTLCassTest(cacheAppOutput, t)
-
-	// be cache service test
-	beCacheServiceOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheservice", []string{"templates/becache-service.yaml"})
-	common.IgniteCacheServiceTest(beCacheServiceOutput, t)
-
-	// configmap test
-	configOutPut := helm.RenderTemplate(t, options, helmChartPath, "configmap", []string{"templates/configmap.yaml"})
-	common.ConfigMapCassandraTest(configOutPut, t)
-}
-
-func TestFTLCacheStoreMysql(t *testing.T) {
-	helmChartPath := "../../"
-	options := &helm.Options{
-		SetValues: common.FTLCacheMysqlStoreValues(),
-	}
-
-	appAndJmxServices(t, options, helmChartPath)
-
-	// inference agent test
-	inferenceOutput := helm.RenderTemplate(t, options, helmChartPath, "beinferenceagent", []string{"templates/beinferenceagent.yaml"})
-	common.InferenceFTLMysqlTest(inferenceOutput, t)
-
-	// cache agent test
-	cacheAppOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheagent", []string{"templates/becacheagent.yaml"})
-	common.CacheFTLMysqlTest(cacheAppOutput, t)
-
-	// be cache service test
-	beCacheServiceOutput := helm.RenderTemplate(t, options, helmChartPath, "becacheservice", []string{"templates/becache-service.yaml"})
-	common.IgniteCacheServiceTest(beCacheServiceOutput, t)
-
-	// configmap test
-	configOutPut := helm.RenderTemplate(t, options, helmChartPath, "configmap", []string{"templates/configmap.yaml"})
-	common.ConfigMapMysqlTest(configOutPut, t)
-}
